refactor: create, update requests with InputObjectType
