--- src/routes/graphql/gqlTypes/index.ts
@@ -1,49 +1,0 @@
-import { GraphQLObjectType, GraphQLString, GraphQLList, GraphQLInt } from 'graphql';
-
-export const UserType = new GraphQLObjectType({
-    name: 'User',
-    fields: () => ({
-      id: { type: GraphQLString },
-      firstName: { type: GraphQLString },
-      lastName: { type: GraphQLString },
-      email: { type: GraphQLString },
-      subscribedToUserIds: { type: new GraphQLList(GraphQLString)},
-      
-      subscriberId: {type: GraphQLString}
-    }),
-  });
-
-export const ProfileType = new GraphQLObjectType({
-    name: 'Profile',
-    fields: () => ({
-      id: { type: GraphQLString },
-      avatar: { type: GraphQLString },
-      sex: { type: GraphQLString },
-      birthday: { type: GraphQLInt },
-      country: { type: GraphQLString },
-      street: { type: GraphQLString },
-      city: { type: GraphQLString },
-      memberTypeId: { type: GraphQLString },
-      userId: { type: GraphQLString },
-    }),
-  });
-
-export const MemberTypeType = new GraphQLObjectType({
-    name: 'MemberType',
-    fields: () => ({
-      id: { type: GraphQLString },
-      discount: { type: GraphQLInt },
-      monthPostsLimit: { type: GraphQLInt },
-    }),
-  });
-
-export const PostType = new GraphQLObjectType({
-    name: 'Post',
-    fields: () => ({
-      id: { type: GraphQLString },
-      title: { type: GraphQLString },
-      content: { type: GraphQLString },
-      userId: { type: GraphQLString },
-    }),
-  });
-
