+++ src/routes/graphql/gqlTypes/profiles.ts
@@ -1,0 +1,44 @@
+import { GraphQLObjectType, GraphQLString,  GraphQLInt, GraphQLInputObjectType, GraphQLNonNull } from 'graphql';
+
+export const ProfileType = new GraphQLObjectType({
+    name: 'Profile',
+    fields: () => ({
+      id: { type: GraphQLString },
+      avatar: { type: GraphQLString },
+      sex: { type: GraphQLString },
+      birthday: { type: GraphQLInt },
+      country: { type: GraphQLString },
+      street: { type: GraphQLString },
+      city: { type: GraphQLString },
+      memberTypeId: { type: GraphQLString },
+      userId: { type: GraphQLString },
+    }),
+});
+
+export const ProfileCreationType = new GraphQLInputObjectType({
+    name: 'ProfileCreation',
+    fields: () => ({
+      avatar: { type: new GraphQLNonNull(GraphQLString) },
+      sex: { type: new GraphQLNonNull(GraphQLString) },
+      birthday: { type: new GraphQLNonNull(GraphQLInt) },
+      country: { type: new GraphQLNonNull(GraphQLString) },
+      street: { type: new GraphQLNonNull(GraphQLString) },
+      city: { type: new GraphQLNonNull(GraphQLString) },
+      memberTypeId: { type: new GraphQLNonNull(GraphQLString) },
+      userId: { type: new GraphQLNonNull(GraphQLString) },
+    }),
+});
+
+export const ProfileUpdatingType = new GraphQLInputObjectType({
+    name: 'ProfileUpdating',
+    fields: () => ({
+      id: { type: new GraphQLNonNull(GraphQLString) },
+      avatar: { type: GraphQLString },
+      sex: { type: GraphQLString },
+      birthday: { type: GraphQLInt },
+      country: { type: GraphQLString },
+      street: { type: GraphQLString },
+      city: { type: GraphQLString },
+      memberTypeId: { type: GraphQLString },
+    }),
+});
