+++ src/routes/graphql/gqlTypes/users.ts
@@ -1,0 +1,50 @@
+import { GraphQLObjectType, GraphQLString, GraphQLList, GraphQLInputObjectType, GraphQLNonNull } from 'graphql';
+
+export const UserType = new GraphQLObjectType({
+    name: 'User',
+    fields: () => ({
+      id: { type: GraphQLString },
+      firstName: { type: GraphQLString },
+      lastName: { type: GraphQLString },
+      email: { type: GraphQLString },
+      subscribedToUserIds: { type: new GraphQLList(GraphQLString)},
+      
+      subscriberId: {type: GraphQLString}
+    }),
+});
+
+export const UserCreationType = new GraphQLInputObjectType({
+    name: 'UserCreation',
+    fields: {
+      firstName: { type: new GraphQLNonNull(GraphQLString) },
+      lastName: { type: new GraphQLNonNull(GraphQLString) },
+      email: { type: new GraphQLNonNull(GraphQLString) },
+    }
+});
+
+export const UserUpdatingType = new GraphQLInputObjectType({
+    name: 'UserUpdating',
+    fields: {
+      id: { type: new GraphQLNonNull(GraphQLString) },
+      firstName: { type: GraphQLString },
+      lastName: { type: GraphQLString },
+      email: { type: GraphQLString },
+      subscribedToUserIds: { type: new GraphQLList(GraphQLString)},
+    }
+});
+
+export const UserSubscribingType = new GraphQLInputObjectType({
+    name: 'UserSubscribing',
+    fields: {
+      user1ID: { type: new GraphQLNonNull(GraphQLString) },
+      user2ID: { type: new GraphQLNonNull(GraphQLString) },
+    }
+});
+
+export const UserUnSubscribingType = new GraphQLInputObjectType({
+    name: 'UserUnSubscribing',
+    fields: {
+      user1ID: { type: new GraphQLNonNull(GraphQLString) },
+      user2ID: { type: new GraphQLNonNull(GraphQLString) },
+    }
+});
